requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
Django==4.1
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
djangorestframework==3.14.0
Faker==20.1.0
flake8==5.0.4
gunicorn==20.1.0
idna==3.4
jarowinkler==1.2.3
Levenshtein==0.20.5
mccabe==0.7.0
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
python-dateutil==2.8.2
pytz==2022.4
rapidfuzz==2.10.2
requests==2.31.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tqdm==4.65.0
typing-extensions==4.8.0
tzdata==2022.1
urllib3==2.0.3
